+++ spec/rubocop/cop/lint/parentheses_as_grouped_expression_spec.rb
@@ -51,7 +51,7 @@ RSpec.describe RuboCop::Cop::Lint::ParenthesesAsGroupedExpression do
     RUBY
   end
 
-  it 'does not register an offense when when method argument parentheses are omitted and ' \
+  it 'does not register an offense when method argument parentheses are omitted and ' \
     'hash argument key is enclosed in parentheses' do
     expect_no_offenses(<<~RUBY)
       transition (foo - bar) => value